refactor: move weight initialization for gpt2 & coca to common parent class
