+++ src/modalities/models/model.py
@@ -1,12 +1,27 @@
+import math
 from abc import abstractmethod
-from typing import Dict, List, Optional
+from enum import Enum
+from functools import partial
+from typing import Annotated, Dict, List, Optional
 
 import torch
 import torch.nn as nn
+from pydantic import BaseModel, Field
 
 from modalities.batch import DatasetBatch, InferenceResultBatch
 
 WeightDecayGroups = Dict[str, List[str]]
+
+
+class ActivationType(str, Enum):
+    GELU = "gelu"
+    FUSED_SWIGLU = "fused_swiglu"
+
+
+class WeightInitializationConfig(BaseModel):
+    mean: Annotated[float, Field(strict=True, ge=0.0)]
+    std: Annotated[float, Field(strict=True, ge=0.0)]
+    type: str
 
 
 class NNModel(nn.Module):
@@ -26,6 +41,25 @@ class NNModel(nn.Module):
 
     def get_parameters(self) -> Dict[str, torch.Tensor]:
         return {name: param for name, param in self.named_parameters()}
+
+    def _init_weights(self, module: nn.Module, weight_init: WeightInitializationConfig):
+        if isinstance(module, nn.Linear):
+            torch.nn.init.normal_(module.weight, mean=weight_init.mean, std=weight_init.std)
+            if module.bias is not None:
+                torch.nn.init.zeros_(module.bias)
+        elif isinstance(module, nn.Embedding):
+            torch.nn.init.normal_(module.weight, mean=weight_init.mean, std=weight_init.std)
+
+    def initialize_weights(self, weight_init, number_of_layers):
+        self.apply(partial(self._init_weights, weight_init=weight_init))
+
+        if weight_init.type == "scaled":
+            # apply special scaled init to the residual projections, per GPT-2 paper
+            for pn, p in self.named_parameters():
+                if pn.endswith("c_proj.weight"):
+                    torch.nn.init.normal_(
+                        p, mean=weight_init.mean, std=weight_init.std / math.sqrt(2 * number_of_layers)
+                    )
 
 
 def model_predict_batch(model: nn.Module, batch: DatasetBatch) -> InferenceResultBatch:
